Implement an interface to access design system UIFonts
