+++ Modules/Sources/DesignSystem/Foundation/TextStyle.swift
@@ -1,4 +1,7 @@
+import struct SwiftUI.Text
+
 public enum TextStyle {
+
     case heading1
     case heading2
     case heading3
@@ -12,5 +15,14 @@ public enum TextStyle {
     public enum Weight {
         case regular
         case emphasized
+    }
+
+    var `case`: Text.Case? {
+        switch self {
+        case .caption:
+            return .uppercase
+        default:
+            return nil
+        }
     }
 }
